🔨 Move `metadata.yaml` related types to `metadata.yaml.ts`
